Removing code duplication (#13)
* release candidate 1.0.0 (#11)

* Swagger enhancement + prepare to release

* release-1.0.0

* rc-1.0.0

* 1.0.0-RC on CHANGELOG

* add context path on landpage redirect

* fix context path on redirect + apix link

* Update README.md

1.0.0-RC

* remove unused lines
