+++ CHANGELOG.md
@@ -1,0 +1,17 @@
+<a name="1.0.0-RC"></a>
+# [1.0.0-RC](https://github.com/groovylabs/lyre/releases/tag/1.0.0-RC) (2017-11-03)
+
+### Features
+
+* **mock endpoints by files:** add Scanner/Reader/Interpreter/APIx classes (lifecycle)
+* **live reload on directory:** add WatchService on Watcher
+* **plug lyre on spring boot application** create @EnableLyre annotation
+* **enable remote connection** customize embedded servlet container configuration
+* **publishing events over websocket** add WebSocketMessageBroker to feed lyre-ui
+* **manage lyre endpoints & configuration using swagger** add Swagger and Management REST services
+* **docker support** add Dockerfile
+
+### Bug Fixes
+
+* **code quality:** work on Reliability/Security/Maintainability via ([Sonarcloud](https://sonarcloud.io/dashboard?id=com.github.groovylabs%3Alyre))
+* **swagger:** project information and useful links. ([8dd3feb](https://github.com/groovylabs/lyre/commit/8dd3feb39fc429f39fe531263180f126c74b7cb1))
