+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.github.groovylabs</groupId>
     <artifactId>lyre</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>Lyre</name>
