+++ sonar-project.properties
@@ -1,6 +1,6 @@
 sonar.projectKey=com.github.groovylabs.lyre
 sonar.projectName=Lyre
-sonar.projectVersion=0.0.1-SNAPSHOT
+sonar.projectVersion=1.0.0-SNAPSHOT
 
 # =====================================================
 #   Properties that will be shared amongst all modules
