+++ src/main/java/com/github/groovylabs/lyre/Lyre.java
@@ -27,28 +27,12 @@ package com.github.groovylabs.lyre;
 
 import com.github.groovylabs.lyre.config.LyreBanner;
 import com.github.groovylabs.lyre.config.LyreProperties;
-import com.github.groovylabs.lyre.engine.APIx.APIx;
-import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.servlet.ServletContainer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jersey.JerseyProperties;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
-import org.springframework.http.HttpStatus;
-import org.springframework.util.StringUtils;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.util.Map;
 
 /**
  * Lyre class - Spring Boot Application responsible to configure and start a sandbox servlet container
@@ -63,68 +47,10 @@ import java.util.Map;
 @EnableConfigurationProperties(LyreProperties.class)
 public class Lyre {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(Lyre.class);
-
-    @Autowired
-    private LyreProperties lyreProperties;
-
     public static void main(String[] args) {
         SpringApplication app = new SpringApplication(Lyre.class);
         app.setMainApplicationClass(Lyre.class);
         app.setBanner(new LyreBanner());
         app.run(args);
     }
-
-    @Bean
-    @Profile("!docker")
-    public EmbeddedServletContainerCustomizer embeddedServletContainerCustomizer() {
-        return (container -> {
-
-            container.setContextPath(lyreProperties.getContextPath());
-
-            if (!lyreProperties.isEnableRemoteConnections()) {
-                try {
-                    InetAddress inetAddress = InetAddress.getByAddress(new byte[]{127, 0, 0, 1});
-                    container.setAddress(inetAddress);
-                } catch (UnknownHostException e) {
-                    //supressed exception
-                }
-            }
-
-            container.setPort(lyreProperties.getPort());
-        });
-    }
-
-    @Bean
-    public ServletRegistrationBean jerseyServletRegistration(
-        JerseyProperties jerseyProperties, ResourceConfig config) {
-
-        ServletRegistrationBean registration = new ServletRegistrationBean(
-            new ServletContainer(config));
-
-        for (Map.Entry<String, String> entry : jerseyProperties.getInit().entrySet()) {
-            registration.addInitParameter(entry.getKey(), entry.getValue());
-        }
-
-        registration.addUrlMappings("/" +
-            (StringUtils.isEmpty(lyreProperties.getApplicationPath()) ? "api" : lyreProperties.getApplicationPath())
-            + "/*");
-        registration.setName(APIx.class.getName());
-        registration.setLoadOnStartup(1);
-        return registration;
-
-    }
-
-    @Bean
-    public WebMvcConfigurerAdapter webMvcConfigurerAdapter() {
-        return new WebMvcConfigurerAdapter() {
-            @Override
-            public void addViewControllers(ViewControllerRegistry registry) {
-                registry.addViewController("/")
-                    .setStatusCode(HttpStatus.PERMANENT_REDIRECT)
-                    .setViewName("/swagger-ui.html");
-            }
-        };
-    }
-
 }
