+++ src/main/java/com/github/groovylabs/lyre/engine/APIx/APIx.java
@@ -96,9 +96,10 @@ public class APIx extends ResourceConfig {
             public void onStartup(final Container container) {
                 LOGGER.info("Lyre REST API Mock tool started");
 
-                LOGGER.info("\u21B3 " + "Endpoints are available at: http://{}:{}",
+                LOGGER.info("\u21B3 " + "Endpoints are available at: http://{}:{}{}",
                     InetAddress.getLoopbackAddress().getHostAddress(),
-                    lyreProperties.getPort());
+                    lyreProperties.getPort(),
+                    lyreProperties.getContextPath());
 
                 com.github.groovylabs.lyre.engine.APIx.APIx.container = container;
             }
